refactor: move inverters to their own file and made approximator docs more general
